requirements.txt
aiogram
python-dotenv
sqlalchemy
pymysql